fix: change biaya column type from mediumInteger to integer in daftar_pemeliharaans migration

+++ database/migrations/2024_11_19_112121_create_daftar_pemeliharaans_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->date('tanggal')->nullable();
             $table->string('uraian')->nullable();
             $table->string('penyedia', 100)->nullable();
-            $table->mediumInteger('biaya')->nullable()->unsigned();
+            $table->integer('biaya')->nullable()->unsigned();
             $table->mediumInteger('pemeliharaan_id')->nullable()->unsigned();
             $table->mediumInteger('master_barang_pemeliharaan_id')->nullable()->unsigned();
             $table->timestamps();
